1. added dummy support for usb transport on FreeBSD. 2. updated get_my_path() functions to be safer and more correct 3. unsigned -> size_t in certain places 4. some code style changes

+++ fastboot/usb_osx.c
@@ -461,7 +461,7 @@ int usb_close(usb_handle *h) {
     return 0;
 }
 
-int usb_read(usb_handle *h, void *data, int len) {
+int usb_read(usb_handle *h, void *data, size_t len) {
     IOReturn result;
     UInt32 numBytes = len;
 
@@ -495,7 +495,7 @@ int usb_read(usb_handle *h, void *data, int len) {
     return -1;
 }
 
-int usb_write(usb_handle *h, const void *data, int len) {
+int usb_write(usb_handle *h, const void *data, size_t len) {
     IOReturn result;
 
     if (len == 0) {
